api interactor tests added plus some refactoring

+++ app/src/main/java/com/winterparadox/themovieapp/personDetails/PersonApiInteractor.java
@@ -6,5 +6,5 @@ import io.reactivex.Single;
 
 public interface PersonApiInteractor {
 
-    Single<Person> personDetails (int id);
+    Single<Person> personDetails (long id);
 }
