+++ app/src/main/java/com/winterparadox/themovieapp/personDetails/PersonApiService.java
@@ -9,5 +9,5 @@ import retrofit2.http.Path;
 public interface PersonApiService {
 
     @GET("person/{id}?append_to_response=movie_credits,tagged_images")
-    Single<Person> personDetails (@Path("id") int personId);
+    Single<Person> personDetails (@Path("id") long personId);
 }
